+++ js/big-pictures.js
@@ -1,5 +1,5 @@
 import {isEscapeKey} from './util.js';
-const maxNumberComment = 5;
+import { createComments} from './createComments.js';
 
 const pictures = document.querySelector('.pictures');
 const bigPicture = document.querySelector('.big-picture');
@@ -8,66 +8,6 @@ const likesCount = document.querySelector('.likes-count');
 const commentsCount = document.querySelector('.comments-count');
 const bigPictureCancel = document.querySelector('.big-picture__cancel');
 const socialCaption = document.querySelector('.social__caption');
-const socialComments = document.querySelector('.social__comments');
-const socialCommentCount = document.querySelector('.social__comment-count');
-const commentsLoader = document.querySelector('.comments-loader');
-const commentTemplate = document.querySelector('#comment').content;
-const documentFragment = document.createDocumentFragment();
-
-const createComments = () => {
-  let commentsNumber = 0;
-  let comments = [];
-  let commentsCountCreate = 0;
-
-  const renderComments = (messageActive) => {
-    socialComments.innerHTML = '';
-
-    messageActive.forEach((comment) => {
-      const commentNode = commentTemplate.cloneNode(true);
-      commentNode.querySelector('.social__picture').src = comment.avatar;
-      commentNode.querySelector('.social__picture').alt = comment.name;
-      commentNode.querySelector('.social__text').textContent = comment.message;
-      documentFragment.append(commentNode);
-    });
-
-    socialComments.append(documentFragment);
-  };
-
-  const eventListener = () => {
-    const newCommentNumber = commentsNumber + maxNumberComment;
-    commentsNumber = newCommentNumber >= commentsCountCreate ? commentsCountCreate : newCommentNumber;
-
-    renderComments(comments.slice(0, commentsNumber));
-
-    socialCommentCount.textContent = `${commentsNumber} из ${commentsCountCreate} комментариев`;
-
-    if (commentsNumber === commentsCountCreate) {
-      commentsLoader.classList.add('hidden');
-    }
-  };
-
-  const addEventListener = () => {
-    commentsLoader.addEventListener('click', eventListener);
-  };
-
-  const removeEventListener = () => {
-    commentsLoader.removeEventListener('click', eventListener);
-  };
-
-  const init = (totalComments) => {
-    commentsCountCreate = totalComments.length;
-    comments = totalComments;
-    commentsNumber = 0;
-    commentsLoader.classList.remove('hidden');
-    eventListener();
-    addEventListener();
-  };
-
-  return {
-    removeEventListener,
-    init
-  };
-};
 
 const commentsCreateFunction = createComments();
 
@@ -82,7 +22,6 @@ const closeModal = (e) => {
 };
 
 const openModal = (photo) => {
-  //let commentsNumberModal = maxNumberComment;
   bigPictureImg.src = photo.url;
   bigPictureImg.alt = photo.description;
   likesCount.textContent = photo.likes;
