+++ js/createComments.js
@@ -1,0 +1,63 @@
+const socialComments = document.querySelector('.social__comments');
+const socialCommentCount = document.querySelector('.social__comment-count');
+const commentsLoader = document.querySelector('.comments-loader');
+const commentTemplate = document.querySelector('#comment').content;
+const documentFragment = document.createDocumentFragment();
+const maxNumberComment = 5;
+const createComments = () => {
+    let commentsNumber = 0;
+    let comments = [];
+    let commentsCountCreate = 0;
+
+    const renderComments = (messageActive) => {
+      socialComments.innerHTML = '';
+
+      messageActive.forEach((comment) => {
+        const commentNode = commentTemplate.cloneNode(true);
+        commentNode.querySelector('.social__picture').src = comment.avatar;
+        commentNode.querySelector('.social__picture').alt = comment.name;
+        commentNode.querySelector('.social__text').textContent = comment.message;
+        documentFragment.append(commentNode);
+      });
+
+      socialComments.append(documentFragment);
+    };
+
+    const eventListener = () => {
+      const newCommentNumber = commentsNumber + maxNumberComment;
+      commentsNumber = newCommentNumber >= commentsCountCreate ? commentsCountCreate : newCommentNumber;
+
+      renderComments(comments.slice(0, commentsNumber));
+
+      socialCommentCount.textContent = `${commentsNumber} из ${commentsCountCreate} комментариев`;
+
+      if (commentsNumber === commentsCountCreate) {
+        commentsLoader.classList.add('hidden');
+      }
+    };
+
+    const addEventListener = () => {
+      commentsLoader.addEventListener('click', eventListener);
+    };
+
+    const removeEventListener = () => {
+      commentsLoader.removeEventListener('click', eventListener);
+    };
+
+    const init = (totalComments) => {
+      commentsNumber = 0;
+      commentsCountCreate = totalComments.length;
+      comments = totalComments;
+      commentsLoader.classList.remove('hidden');
+      eventListener();
+      addEventListener();
+    };
+
+    return {
+      removeEventListener,
+      init
+    };
+  };
+export {
+  createComments
+};
